refactor: namespace private native and test filenames
Rename private native sources and headers to use the mpq- namespace.
Rename test sources and helpers to use the test-mpq- namespace while
preserving test executable names.

Update Automake lists, includes, packaging references, and source checks.
Keep the public API, ABI, library name, and installed headers unchanged.
