+++ src/mpq-huffman.h
@@ -1,9 +1,9 @@
 /*
- *  huffman.h -- adaptive Huffman decompression structures and helpers.
+ *  mpq-huffman.h -- adaptive Huffman decompression structures and helpers.
  *
  *  Copyright (c) 2003-2026 Maik Broemme <mbroemme@libmpq.org>
  *
- *  This source was adapted from the C++ version of huffman.h included
+ *  This source was adapted from the C++ version of mpq-huffman.h included
  *  in stormlib. The C++ version belongs to the following authors:
  *
  *  Ladislav Zezula <ladik@zezula.net>
