+++ src/mpq-wave.h
@@ -1,9 +1,9 @@
 /*
- *  wave.h -- MPQ WAVE decompression declarations and lookup tables.
+ *  mpq-wave.h -- MPQ WAVE decompression declarations and lookup tables.
  *
  *  Copyright (c) 2003-2026 Maik Broemme <mbroemme@libmpq.org>
  *
- *  This source was adapted from the C++ version of wave.h included
+ *  This source was adapted from the C++ version of mpq-wave.h included
  *  in stormlib. The C++ version belongs to the following authors:
  *
  *  Ladislav Zezula <ladik.zezula.net>
