+++ src/mpq-writer.c
@@ -1,5 +1,5 @@
 /*
- *  writer.c -- seekable MPQ archive creation.
+ *  mpq-writer.c -- seekable MPQ archive creation.
  *
  *  Copyright (c) 2003-2026 Maik Broemme <mbroemme@libmpq.org>
  *
@@ -20,13 +20,13 @@
 #include "config.h"
 #endif
 
-#include "compression.h"
-#include "crypto.h"
-#include "endian.h"
+#include "mpq-compression.h"
+#include "mpq-crypto.h"
+#include "mpq-endian.h"
 #include "mpq-internal.h"
-#include "pkware.h"
-#include "wave.h"
-#include "writer.h"
+#include "mpq-pkware.h"
+#include "mpq-wave.h"
+#include "mpq-writer.h"
 #include <libmpq/mpq.h>
 
 #include <stdio.h>
